refactor: eliminate duplicated types in web-ui by importing from shared schema
Web-ui types files (triage.ts, snapshot.ts) manually mirrored types from
src/schema/ and src/export/types.ts, risking silent divergence. Now all
types are defined in @kynetic-ai/shared and web-ui re-exports from there.

Changes:
- Add Convention, TriageStatus, TriageAction, TriageRecord, and export
  snapshot types (ExportedTask, ExportedItem, InheritedAC, ExportedValidation,
  KspecSnapshot, ExportedProject) to @kynetic-ai/shared api.ts
- Replace web-ui types/triage.ts with re-export from shared
- Replace web-ui types/snapshot.ts with re-exports from shared
- Add static analysis test enforcing no local type definitions in types dir
- Update web-ui-triage test to verify re-export pattern

Task: @task-deduplicate-web-ui-types
